+++ MainServer/include/Structures/Capsule/CapsuleList.h
@@ -14,8 +14,8 @@ PACK_PUSH(1)
 			std::uint32_t capsuleInfoId = 0;
 			std::uint32_t newPrice : 21 = 0;
 			std::uint32_t unknown2 : 11 = 0;
-			std::uint32_t saleEventStartDate{};
-			std::uint32_t saleEventEndDate{};
+			std::uint64_t saleEventStartDate{};
+			std::uint64_t saleEventEndDate{};
 		};
 PACK_POP()
 
@@ -24,24 +24,24 @@ PACK_PUSH(1)
 		{
 			std::uint32_t newRtPrice{};
 			std::uint32_t newMpPrice{};
-			std::uint32_t saleEventStartDate{};
-			std::uint32_t saleEventEndDate{};
+			std::uint64_t saleEventStartDate{};
+			std::uint64_t saleEventEndDate{};
 		};
 PACK_POP()
 
 PACK_PUSH(1)
 		struct EventMissionInfo
 		{
-			std::uint32_t startDate{};
-			std::uint32_t endDate{};
+			std::uint64_t startDate{};
+			std::uint64_t endDate{};
 		};
 PACK_POP()
 
 PACK_PUSH(1)
 		struct ExpMpBonusInfo
 		{
-			std::uint32_t startDate{};
-			std::uint32_t endDate{};
+			std::uint64_t startDate{};
+			std::uint64_t endDate{};
 			std::uint32_t expBonusPercent{};
 			std::uint32_t mpBonusPercent{};
 		};
